Let `OverlayView` be clusterd in `MarkerClusterer`
Since `OverlayView` can represent any kind of overlay, including extremely
custom markers, it makes sense to allow it to be clustered. For this to work,
we only have to implement two functions of the `Marker` class: `getDraggable`
and `getPosition`.

The `MarkerClusterer` will call our `OverlayView`'s `setVisible` method as
needed.